(!2007) dix: unexport and document MaybeDeliverEventsToClient()
Not used by any external drivers, and only supposed to be DIX internal,
so shouldn't be exported at all.

+++ dix/events.c
@@ -2547,7 +2547,7 @@ XineramaTryClientEventsResult(ClientPtr client,
  */
 int
 MaybeDeliverEventsToClient(WindowPtr pWin, xEvent *pEvents,
-                           int count, Mask filter, ClientPtr dontClient)
+                           size_t count, Mask filter, ClientPtr dontClient)
 {
     OtherClients *other;
